requirements.txt
Appium-Python-Client==2.7.0
async-generator==1.10
attrs==22.1.0
certifi==2022.9.24
exceptiongroup==1.0.0rc9
future==0.18.2
h11==0.14.0
idna==3.4
iniconfig==1.1.1
iso8601==1.1.0
loguru==0.6.0
outcome==1.2.0
packaging==21.3
pluggy==0.13.1
pure-python-adb==0.3.0.dev0
py==1.11.0
pyparsing==3.0.9
pyserial==3.5
PySocks==1.7.1
pytest==6.1.1
PyYAML==6.0
selenium==4.5.0
serial==0.0.97
sniffio==1.3.0
sortedcontainers==2.4.0
toml==0.10.2
trio==0.22.0
trio-websocket==0.9.2
typing==3.7.4.3
urllib3==1.26.12
wsproto==1.2.0
